requirements.txt
flask
flask_cors
flask_sqlalchemy
google-generativeai
python-dotenv